chore: remove component menu v2
